open-ai-agent-sdk/5.streaming/pyproject.toml
[project]
name = "5-streaming"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "openai>=1.97.1",
    "openai-agents>=0.2.3",
    "python-dotenv>=1.1.1",
]
